Refactor session management and add session validation for APIs
- Moved session-related utility functions (`createSignedCookie`, `validateSessionCookie`) to a shared module in `utils/session.js`.
- Updated `auth.js` to use the shared `createSignedCookie` function for session cookie creation.
- Added session cookie validation in `generate-image.js` and `translate.js` to ensure authorized access.
- Implemented Origin validation in `generate-image.js` and `translate.js` to restrict requests to allowed origins.
- Improved code readability by centralizing session management logic and reducing duplication.
