+++ functions/generate-image.js
@@ -1,4 +1,16 @@
+import { validateSessionCookie } from "../utils/session.js";
+
 export async function onRequestPost({ request, env }) {
+  const sessionSecret = new TextEncoder().encode(env.SESSION_SECRET);
+
+  // セッションCookieの検証
+  const isValidSession = await validateSessionCookie(request, sessionSecret);
+  if (!isValidSession) {
+    console.error("認証エラー: 無効なセッションCookie");
+    return new Response("Unauthorized", { status: 401 });
+  }
+
+  // Origin検証
   const allowedOrigin = env.ALLOWED_ORIGIN;
   const origin = request.headers.get("Origin");
   if (allowedOrigin !== origin) {
@@ -7,6 +19,8 @@ export async function onRequestPost({ request, env }) {
     );
     return new Response("Forbidden", { status: 403 });
   }
+
+  // 画像生成の処理
   const formData = await request.formData();
   const prompt = formData.get("prompt");
   return await generateImage(prompt, env);
